review comments: add explicit typenames

+++ contracts/forwarder/Forwarder.sol
@@ -59,15 +59,13 @@ contract Forwarder is IForwarder {
         _verifyAndUpdateNonce(req);
 
         bytes memory callData = abi.encodePacked(req.data, req.from);
-        uint preBalance = address (this).balance;
+        uint256 preBalance = address(this).balance;
         require(preBalance >= req.value, "FWD: insufficient value");
-        require( gasleft()*63/64 >= req.gas, "FWD: insufficient gas" );
+        require(gasleft()*63/64 >= req.gas, "FWD: insufficient gas" );
         // solhint-disable-next-line avoid-low-level-calls
-        (success,ret) = req.to.call{gas : req.gas, value : req.value}(callData);
+        (success,ret) = req.to.call{gas: req.gas, value: req.value}(callData);
 
-        //this is a signed number (can be nagative). all params are known values,
-        // so can't overflow.
-        int excessBalance = int(address (this).balance - preBalance + req.value);
+        int256 excessBalance = int256(address(this).balance - preBalance + req.value);
 
         // preBalance is the balance before this TX (and value sent to it as payable)
         // preBalance-req.value - the expected balance after the call, assuming it didn't send anything back
@@ -77,7 +75,7 @@ contract Forwarder is IForwarder {
         //         this excess is inaccessible.
         if ( excessBalance>0 ) {
             //can't fail: req.from signed (off-chain) the request, so it must be an EOA...
-            payable(req.from).transfer(uint(excessBalance));
+            payable(req.from).transfer(uint256(excessBalance));
         }
         return (success,ret);
     }
